move utils to components dir. refs #99
