+++ components/ui/avatar.tsx
@@ -2,8 +2,7 @@
 
 import * as AvatarPrimitive from "@radix-ui/react-avatar";
 import * as React from "react";
-
-import { cn } from "@/lib/utils";
+import { cn } from "../utils";
 
 const Avatar = React.forwardRef<
   React.ElementRef<typeof AvatarPrimitive.Root>,
