+++ components/ui/menubar.tsx
@@ -3,8 +3,7 @@
 import * as MenubarPrimitive from "@radix-ui/react-menubar";
 import { Check, ChevronRight, Circle } from "lucide-react";
 import * as React from "react";
-
-import { cn } from "@/lib/utils";
+import { cn } from "../utils";
 
 const MenubarMenu = MenubarPrimitive.Menu;
