+++ components/ui/resizable.tsx
@@ -1,7 +1,6 @@
 import { GripVertical } from "lucide-react";
 import * as ResizablePrimitive from "react-resizable-panels";
-
-import { cn } from "@/lib/utils";
+import { cn } from "../utils";
 
 const ResizablePanelGroup = ({
   className,
