+++ components/ui/sheet.tsx
@@ -4,8 +4,7 @@ import * as React from "react";
 import * as SheetPrimitive from "@radix-ui/react-dialog";
 import { cva, type VariantProps } from "class-variance-authority";
 import { X } from "lucide-react";
-
-import { cn } from "@/lib/utils";
+import { cn } from "../utils";
 
 const Sheet = SheetPrimitive.Root;
