+++ components/ui/tabs.tsx
@@ -2,8 +2,7 @@
 
 import * as TabsPrimitive from "@radix-ui/react-tabs";
 import * as React from "react";
-
-import { cn } from "@/lib/utils";
+import { cn } from "../utils";
 
 const Tabs = TabsPrimitive.Root;
